requirements.txt
# Function dependencies, for example:
# package>=version

google-auth
google-cloud-storage
google-cloud-bigquery
datetime
requests
pandas
numpy
google-cloud-secret-manager==2.0.0
